DataFlow To StudyIDs Completed
Revised ISQLColumn, ISQLTable, ISQLDatabase ISQLServer interfaces, unifying implementation and hopefully simplyfing linq quries based on them.
Revisited, and updated all related data fetching classes and interfaces focused around the above change.

DataContext now successfully returning studyids.

Simplified creation of IEnumerableDataSource implementers by adding AIEnumerableDataSource abstract base class for common functionality.

ISQLTableColumnValues has been replaced by ISQLTableColumn.

SQLTableColumnValues.cs has been replaced by SQLTableColumn.

+++ ScriptGeneratorRedux/Models/Core/IO/Interfaces/IDataContext.cs
@@ -13,8 +13,8 @@ namespace ScriptGeneratorRedux.Models.Core.IO.Interfaces
         IEnumerable<String> GetEnvironmentNames( String ServerName = null );
         IEnumerable<String> GetSecurityDBNames( String ServerName = null );
         IEnumerable<String> GetServerNames( );
-        IEnumerable<Int64> GetStudyIDs( String ServerName, ECP4DepoplymentEnvironment EnvironmentName );
-        IEnumerable<Int64> GetStudyIDs( String ServerName = null );
+        IEnumerable<int> GetStudyIDs( String ServerName, ECP4DepoplymentEnvironment EnvironmentName );
+        IEnumerable<int> GetStudyIDs( String ServerName = null );
         void RegisterServerDetailsProvider( ISQLServerProvider SQLServerProvider );
         void UpdateServersList( );
     }
